requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudflare==2.9.10
flake8==4.0.1
idna==3.3
jsonlines==3.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
yapf==0.32.0
